chore: separate cli-entry and cli tests (#6489)
This is cleanup from #6479. Everything is being tested the
same as before, but this fixes the EventEmitter max listeners
warning that was happening because exit-handler was not being
mocked properly
